% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_calc_wt.R
\name{add_calc_wt}
\alias{add_calc_wt}
\title{Add Calculated Weight}
\usage{
add_calc_wt(x, stock, units = "t")
}
\arguments{
\item{x}{crab data that includes fields 'size', 'sex', and 'clutch' and 'maturity' for chionoecetes stocks (if use_observer = T) and 'crab_year' (if use_historic = T).}

\item{stock}{NULL. Character string stock abbreviation: BSSC, WBT, EBT, BBRKC, EAG, WAG, PIGKC, SMBKC, PIBKC, PIRKC, WAIRKC.}

\item{units}{NULL. Unit of measurement, "t" or "lb". Default = "t".}
}
\value{
x with additional column 'calc_wt'
}
\description{
Add calculated weight based on weight at length parameters.
}
\examples{
add_calc_wt(data, stock = "BBRKC")

}
