% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_crab_year.R
\name{add_crab_year}
\alias{add_crab_year}
\title{Add crab year to data}
\usage{
add_crab_year(x, date_correct = T, date_format = "mdy")
}
\arguments{
\item{x}{Observer or dockside data frame with "fishery" code column.}

\item{date_correct}{Do crab year correction if date is in different crab year than regulatory year at the time, default = T.}

\item{date_format}{Format of date field, default = "mdy".}
}
\value{
x with additional column 'crab_year'
}
\description{
Add crab year to observer and dockside data
}
\examples{
add_crab_year(dockside)
}
