% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_permit_holder.R
\name{add_permit_holder}
\alias{add_permit_holder}
\title{Add Permit Holder to Observer  Pot Data}
\usage{
add_permit_holder(ft_dump, obs, permit_path = "aigkc/data/permits")
}
\arguments{
\item{ft_dump}{NULL. AIGKC fish ticket dump.}

\item{pot_data}{NULL. Observer count pot data in the same format as the output of \code{load_pot_dump()}}

\item{permits_path}{NULL. File path to CFEC permit data directory}
}
\value{
Pot data with additional column 'permit_holder'.
}
\description{
Use fish ticket dump and CFEC permit holder information to join permit holder as proxy for captain to AIGKC data
}
\examples{
load_pot_dump("./data.csv", stock = "BBRKC")

}
