% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_target_stock.R
\name{add_target_stock}
\alias{add_target_stock}
\title{Add target stock to data}
\usage{
add_target_stock(x, col = fishery)
}
\arguments{
\item{x}{data frame}

\item{col}{fishery code colum, default = fishery}
}
\value{
x with additional column 'target_stock'
}
\description{
Add target stock based on fishery code
}
\examples{
add_target_stock(x)
}
