% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_lat_lon.R
\name{clean_lat_lon}
\alias{clean_lat_lon}
\title{Clean Lat and Lon}
\usage{
clean_lat_lon(x)
}
\arguments{
\item{x}{Pot or observer data with columns longitude, latitude, and eastwest}
}
\value{
x with location information correctect across the date line to be negative
}
\description{
Clean latitude and longitude data across the dateline
}
\examples{
clean_lat_lon(pots)
}
