% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockside_comp.R
\name{dockside_comp}
\alias{dockside_comp}
\title{Get retained size composition}
\usage{
dockside_comp(data, by, lump = T)
}
\arguments{
\item{data}{NULL. Dockside sample data in the format as the output of \code{load_dockside()}}

\item{by}{NULL. Vector of column names other than crab year, fishery, and size to use as a grouping variable.}

\item{lump}{T/F lump shell condition into new and old (TRUE) or not (FALSE). Default = TRUE.}
}
\value{
Size frequency by year, fishery, and size
}
\description{
Compute retained size composition from dockside data
}
\examples{
dockside_comp(data, by = "shell", lump = T)

}
