% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discards.R
\name{get_discards}
\alias{get_discards}
\title{Get Directed Discards}
\usage{
get_discards(retained_catch, total_catch, stock, handling_mortality = 0.3)
}
\arguments{
\item{retained_catch}{NULL. Output of \code{get_retaied_catch()}.}

\item{total_catch}{NULL. Output of \code{get_total_catch()}. 'by' argument could be anything as long as group or sex is included.}

\item{stock}{NULL. Character string stock abbreviation: BSSC, WBT, EBT, BBRKC, EAG, WAG, PIGKC, SMBKC, PIBKC, PIRKC, WAIRKC.}
}
\value{
Data frame including crab year, fishery, sex, discards, and discard mortality.
}
\description{
Estimate directed fishery discards by sex using subtraction method
}
\examples{
get_discards(retained_catch, total_catch, stock = "BBRKC", handling_mortality = 0.3)

}
