% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observer_comp.R
\name{get_observer_comp}
\alias{get_observer_comp}
\title{Get observer size composition}
\usage{
get_observer_comp(data, by = NULL, lump = T)
}
\arguments{
\item{data}{NULL. Dockside sample data in the format as the output of \code{load_crab_dump()}}

\item{by}{NULL. Vector of column names other than crab year, fishery, and size to use as a grouping variable.}

\item{lump}{T/F. Lump shell condition into new and old (TRUE) or not (FALSE). Default = TRUE.}
}
\value{
Size frequency by year, fishery, and size
}
\description{
Compute size composition from observer measure pot data
}
\examples{
dockside_comp(data, by = "shell", lump = T)

}
