% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_retained_catch.R
\name{get_retained_catch}
\alias{get_retained_catch}
\title{Get Retained Catch}
\usage{
get_retained_catch(ft_data, by = NULL, units = "t")
}
\arguments{
\item{ft_data}{NULL. Fish ticket data with fields 'crab_year', 'fishery', 'effort' and whatever variables are specified in by.}

\item{by}{NULL. Vector of column names other than crab year, fishery to use as a grouping variable. Must by present in ft_data.}

\item{units}{NULL. Unit of measurement, "t" or "lb". Default = "t".}
}
\value{
Data frame including crab year, fishery, grouping variables, and average weight.
}
\description{
Get retained catch from fish ticket data
}
\examples{
get_retained_catch(ft_data, by = NULL)

}
