% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pot_dump.R
\name{load_pot_dump}
\alias{load_pot_dump}
\title{Load BSAI Observer Count Pot Data}
\usage{
load_pot_dump(path, stock, database_pull = F, clean = T)
}
\arguments{
\item{path}{NULL. Path to data file if not pulling directly from database.}

\item{stock}{NULL. Character string stock abbreviation: BSSC, WBT, EBT, BBRKC, EAG, WAG, PIGKC, SMBKC, PIBKC, PIRKC, WAIRKC.}

\item{database_pull}{Default = FALSE. Do updated pull from database.}

\item{clean}{Default = TRUE. Do stock specific data cleaning.}
}
\value{
Measure pot data time series by fishery.
}
\description{
Load BSAI observer count pot data and do data management routine
}
\examples{
load_pot_dump("./data.csv", stock = "BBRKC")

}
