% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observer_comp.R
\name{get_avg_wt}
\alias{get_avg_wt}
\title{Get Average Weight}
\usage{
get_avg_wt(data, by, lump = T)
}
\arguments{
\item{data}{NULL. Dockside sample data in the format as the output of \code{load_crab_dump()}}

\item{by}{NULL. Vector of column names other than crab year, fishery, and size to use as a grouping variable.}

\item{lump}{T/F Lump shell condition into new and old (TRUE) or not (FALSE). Default = TRUE.}

\item{units}{NULL. Unit of measurement, "t" or "lb". Default = "t".}
}
\value{
Data frame including crab year, fishery, grouping variables, and average weight.
}
\description{
Get average weight of crab by grouping variable using weight at length parameters
}
\examples{
get_avg_wt(data, by = "group")

}
